requirements.txt
Flask==2.1.1
gevent==21.8.0
requests==2.26.0
Pillow==8.4.0
numpy==1.21.4
Keras==2.6.0
tensorflow==2.7.0  # Adjust the version depending on your TensorFlow installation
